Improved rating system.

+++ db/__init__.py
@@ -21,7 +21,7 @@ class Speeders(Base):
     uniqueID = Column(String, index=True)
     datetime = Column(DateTime, index=True)
     speed = Column(Float, index=True)
-    rating = Column(Integer)
+    rating = Column(Float)
 
 class Vehicles(Base):
     """
